Don't use the JMH CLI to execute JMH
So far the plugin used the JMH CLI options to execute benchmarks. The problem
is that the CLI options are not stable. This commit changes the implementation
so that we use the JMH API instead. This has several consequences:

- the plugin now builds using its own version of JMH
- a `runner.jar` is embedded and contains a minimal launcher for JMH benchmarks
- when the benchmarks are executed, we use the version provided by the user

In other words, to be able to use the JMH API, we're compiling against a version
of JMH which might not be the one that the user will use to compile the benchmarks
and execute them.

In general, this shouldn't be a problem, because the API is supposed to be stable.