Make the JMH runtime configuration a "hard" configuration
This fixes #108, where dependency resolution rules were not applied to the JMH runtime.
This is caused by Gradle not applying resolution rules on detached configurations. While a
detached configuration makes sense here (because we don't want the user to manipulate the
JMH runtime, they must never have access to it), in practice it breaks some user builds.

This is a temporary fix until Gradle provides an API to apply resolution rules on detached
configurations.