Rework the plugin implementation
This commit revamps the implementation of the JMH plugin. First of all,
this rolls back to using the CLI to execute JMH: using the API proved
to be a bad idea in practice, due to many different issues (from classloading
to locking or memory leaks).

Second, it fully embraces the Gradle lazy configuration API.

To recognize those changes, the plugin id was changed to `me.champeau.jmh`.
The packages also changed to `me.champeau.jmh`.

A new test was added to demonstrate how to use Java feature previews.
The minimal Gradle version supported is Gradle 6.8.