Modify Kotlin functional test to use internal code
Kotlin has the concept of "internal" code - code which is only accessible to the current "module". When running JMH benchmarks, I believe most developers would consider their benchmarks to be part of the same module. Right now, the JMH plugin does not resolve internal APIs from the main sourceset to the jmh sourceset.