Associate the JMH Kotlin compilation with the main compilation
This allows Kotlin developers to use "internal" code in their benchmarks.

I modified the test in 2 ways. Firstly, I switched to applying the Kotlin plugin via the plugins API instead of "apply". When using "apply", the plugin is not resolved early enough. Given that the "apply" method is legacy, and this is new functionality, it seems like a reasonable requirement for JMH plugin users to have to apply Kotlin via the DSL. Secondly, I changed the version to 1.3.60. This was the first version of the Kotlin plugin that had the required APIs. See https://youtrack.jetbrains.com/issue/KT-34901 for details.

Fixes #197