Allows `jmh` definition to specifies environment variables
Sometimes it is necessary to pass some environment variables to configure
the JVM process (in particular when native libraries are involved). And
when there's no system property equivalent, or when it's not possible to do in the benchmark setup (e.g. `LD_PRELOAD`, `JAVA_LIBRARY_PATH`).

Alas, JMH does not support to set environment variable on the forked
process, however it will propagate the ones that it has been launched with.
E.g.

```
env JAVA_LIBRARY_PATH=/somewhere java -jar bench.jar
```

This change aims at providing the ability to pass the variable to JMH so it cna be used on the JMH forked process.